[US-0004] - Content Creator Login #53
* new route

Added a new rotue with is for login verification for content-creators

* login system

Login system no cryptation

* tab

* email verfication resposes

changed some errors

* fixed a bug using test

Created a http request to test the system
found a mistake and fixed it

* Backend works as it should

TO run the http Test you need to create a user with the email and password depictaed in the http

* Revert "Merge branch 'dev' into feature-ContentCreatorLogin"

This reverts commit 823cfd45de2ca5eba320e976983f490d38bf20dd, reversing
changes made to 003d93225fb3869410396ebe80822ea443202e7d.

* causel

causel fixe ui

* fixed lendning

* more eslint

* welcome

* Texst placed the right place

* documents

* update complete

* it works?

* begining to do stuff

* feedback from luise

* error handling

* Spelling mistake fix

* Tailwind login update

Added tailwind colors and fonts to login but some of the colors are wrong  so that has to be fixed.

* npm run

* new route

Added a new rotue with is for login verification for content-creators

* login system

Login system no cryptation

* tab

* email verfication resposes

changed some errors

* fixed a bug using test

Created a http request to test the system
found a mistake and fixed it

* Backend works as it should

TO run the http Test you need to create a user with the email and password depictaed in the http

* Revert "Merge branch 'dev' into feature-ContentCreatorLogin"

This reverts commit 823cfd45de2ca5eba320e976983f490d38bf20dd, reversing
changes made to 003d93225fb3869410396ebe80822ea443202e7d.

* causel

causel fixe ui

* fixed lendning

* more eslint

* welcome

* Texst placed the right place

* documents

* update complete

* it works?

* begining to do stuff

* feedback from luise

* error handling

* Spelling mistake fix

* Tailwind login update

Added tailwind colors and fonts to login but some of the colors are wrong  so that has to be fixed.

* npm run

* Restore missing files from origin/dev

* Foo

* content creator stuff

* no white screen

* it works

* last bit changes

* eslist

* figma things that we forgot

* test and found an error

* some add notes

* add co uthor

Co-Authored-By: Sozh03 <152170949+Sozh03@users.noreply.github.com>

* Delete strapi/src/api/content-creator/content-creator.http

* errorCodes.ts

in responds to comment by Cathredral101

Co-Authored-By: Sozh03 <152170949+Sozh03@users.noreply.github.com>

* jwt key

update based on comment

Co-Authored-By: Sozh03 <152170949+Sozh03@users.noreply.github.com>

* Update content-creator.ts

Co-Authored-By: Sozh03 <152170949+Sozh03@users.noreply.github.com>

* thisngs that some how got deleted

Co-Authored-By: Sozh03 <152170949+Sozh03@users.noreply.github.com>

* shouldn't have been changed

Co-Authored-By: Sozh03 <152170949+Sozh03@users.noreply.github.com>

* deleted per request

* type check content creator

---------

Co-authored-by: Sooonnnaaa <Bubulspam@gmail.com>
Co-authored-by: Oliver Roed Schøler <oliver@roedscholer.dk>
Co-authored-by: Sozh03 <152170949+Sozh03@users.noreply.github.com>
Co-authored-by: Johny1511 <83210690+Johny1511@users.noreply.github.com>